+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 		EnableHTTPs("cert.pem", "key.pem").
 		Launch()
 	```
+- fix(#183): reject void function as handler
 - add `OnLaunch` fairing
 	```go
 	type YourFairing struct {