+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Latest
 
+## v0.14.0
+
 - (#165) add directly access to `*http.Request`
 	```go
 	type RequestContext struct {